pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "trading_bot"
version = "0.1.0"
dependencies = [
    "tastytrade",
    "gspread",
    "google-auth",
    "pydantic",
    "pyyaml",
    "scipy",
    # add others
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["trading_bot"]